test/fix: changed calculate_area to correctly handle multipolygons
