+++ tests/fixtures/feature-multipolygon-germany-heidelberg.geojson
@@ -1,0 +1,57 @@
+ {
+      "type": "Feature",
+      "properties": {},
+      "geometry": {
+        "coordinates": [
+          [
+            [
+              [
+                8.661708721986429,
+                49.42101485649272
+              ],
+              [
+                8.661708721986429,
+                49.412389754016516
+              ],
+              [
+                8.676124506805195,
+                49.412389754016516
+              ],
+              [
+                8.676124506805195,
+                49.42101485649272
+              ],
+              [
+                8.661708721986429,
+                49.42101485649272
+              ]
+            ]
+          ],
+          [
+            [
+              [
+                8.678853307373373,
+                49.421050716540265
+              ],
+              [
+                8.678853307373373,
+                49.41274841633887
+              ],
+              [
+                8.690678109833613,
+                49.41274841633887
+              ],
+              [
+                8.690678109833613,
+                49.421050716540265
+              ],
+              [
+                8.678853307373373,
+                49.421050716540265
+              ]
+            ]
+          ]
+        ],
+        "type": "MultiPolygon"
+      }
+    }
